fix(VisitorsList): use separate stomp instance.
